extract JDK check logic from SSHLauncher, remove unused code, fix tests
