+++ src/main/java/hudson/plugins/sshslaves/SSHLauncher.java
@@ -47,17 +47,14 @@ import com.trilead.ssh2.SFTPv3Client;
 import com.trilead.ssh2.SFTPv3FileAttributes;
 import com.trilead.ssh2.ServerHostKeyVerifier;
 import com.trilead.ssh2.Session;
-import com.trilead.ssh2.transport.TransportManager;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import hudson.AbortException;
 import hudson.EnvVars;
 import hudson.Extension;
-import hudson.ExtensionList;
 import hudson.Util;
 import hudson.model.Descriptor;
 import hudson.model.Hudson;
 import hudson.model.ItemGroup;
-import hudson.model.JDK;
 import hudson.model.Node;
 import hudson.model.Slave;
 import hudson.model.TaskListener;
@@ -71,10 +68,6 @@ import hudson.slaves.NodeProperty;
 import hudson.slaves.NodePropertyDescriptor;
 import hudson.slaves.SlaveComputer;
 import hudson.tools.JDKInstaller;
-import hudson.tools.JDKInstaller.CPU;
-import hudson.tools.JDKInstaller.Platform;
-import hudson.tools.ToolLocationNodeProperty;
-import hudson.tools.ToolLocationNodeProperty.ToolLocation;
 import hudson.util.DescribableList;
 import hudson.util.FormValidation;
 import hudson.util.ListBoxModel;
@@ -87,7 +80,6 @@ import org.acegisecurity.context.SecurityContext;
 import org.acegisecurity.context.SecurityContextHolder;
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.output.TeeOutputStream;
 import org.apache.commons.lang.StringUtils;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
@@ -96,28 +88,17 @@ import org.kohsuke.stapler.AncestorInPath;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.QueryParameter;
 
-import java.io.BufferedOutputStream;
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.PrintStream;
-import java.io.StringWriter;
 import java.lang.InterruptedException;
-import java.lang.reflect.Field;
-import java.net.URL;
 import java.text.MessageFormat;
-import java.text.NumberFormat;
-import java.text.ParseException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Set;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
@@ -128,7 +109,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.regex.Pattern;
 
 import static com.cloudbees.plugins.credentials.CredentialsMatchers.*;
 import com.cloudbees.plugins.credentials.common.StandardUsernameListBoxModel;
@@ -136,12 +116,10 @@ import edu.umd.cs.findbugs.annotations.CheckForNull;
 import static hudson.Util.*;
 import hudson.model.Computer;
 import hudson.security.AccessControlled;
-import hudson.util.VersionNumber;
 
 import javax.annotation.Nonnull;
 import java.io.UnsupportedEncodingException;
 import java.nio.charset.Charset;
-import static java.util.logging.Level.*;
 
 /**
  * A computer launcher that tries to start a linux slave by opening an SSH connection and trying to find java.
@@ -153,10 +131,6 @@ public class SSHLauncher extends ComputerLauncher {
      */
     public static final SchemeRequirement SSH_SCHEME = new SchemeRequirement("ssh");
 
-
-    public static final String JDKVERSION = "jdk-8u144";
-    public static final String DEFAULT_JDK = JDKVERSION + "-oth-JPR";
-
     // Some of the messages observed in the wild:
     // "Connection refused (Connection refused)"
     // "Connection reset"
@@ -166,20 +140,6 @@ public class SSHLauncher extends ComputerLauncher {
     private static final List<String> RECOVERABLE_FAILURES = Arrays.asList(
             "Connection refused", "Connection reset", "Connection timed out", "No route to host", "Premature connection close"
     );
-
-    /**
-     * @deprecated
-     *      Subtype of {@link JDKInstaller} causes JENKINS-10641.
-     */
-    public static class DefaultJDKInstaller extends JDKInstaller {
-        public DefaultJDKInstaller() {
-            super(DEFAULT_JDK, true);
-        }
-
-        public Object readResolve() {
-            return new JDKInstaller(DEFAULT_JDK,true);
-        }
-    }
 
     /**
      * Field host
@@ -231,14 +191,6 @@ public class SSHLauncher extends ComputerLauncher {
      * Field javaPath.
      */
     public final String javaPath;
-
-    /**
-     * to install JDK, keep this field null. This avoids baking the default value into the persisted form.
-     * @see #getJDKInstaller()
-     *
-     */
-    @Deprecated
-    private JDKInstaller jdk = null;
 
     /**
      * SSH connection to the slave.
@@ -303,7 +255,7 @@ public class SSHLauncher extends ComputerLauncher {
      * @param port       The port to connect on.
      * @param credentialsId The credentials id to connect as.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      * @param launchTimeoutSeconds Launch timeout in seconds
@@ -325,7 +277,7 @@ public class SSHLauncher extends ComputerLauncher {
      * @param port       The port to connect on.
      * @param credentialsId The credentials id to connect as.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      * @param launchTimeoutSeconds Launch timeout in seconds
@@ -383,7 +335,7 @@ public class SSHLauncher extends ComputerLauncher {
      * @param port       The port to connect on.
      * @param credentials The credentials to connect as.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      * @param launchTimeoutSeconds Launch timeout in seconds
@@ -427,7 +379,7 @@ public class SSHLauncher extends ComputerLauncher {
      * @param password   The password to connect with.
      * @param privatekey The ssh privatekey to connect with.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      * @deprecated use the {@link StandardUsernameCredentials} based version
@@ -448,8 +400,8 @@ public class SSHLauncher extends ComputerLauncher {
      * @param password   The password to connect with.
      * @param privatekey The ssh privatekey to connect with.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
-     * @param jdkInstaller The jdk installer that will be used if no java vm is found on the specified host. If <code>null</code> the {@link DefaultJDKInstaller} will be used.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected.
+     * @param jdkInstaller not used.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      * @deprecated use the {@link StandardUsernameCredentials} based version
@@ -466,9 +418,6 @@ public class SSHLauncher extends ComputerLauncher {
         this.credentials = null;
         this.credentialsId = null;
         this.javaPath = fixEmpty(javaPath);
-        if (jdkInstaller != null) {
-            this.jdk = jdkInstaller;
-        }
         this.prefixStartSlaveCmd = fixEmpty(prefixStartSlaveCmd);
         this.suffixStartSlaveCmd = fixEmpty(suffixStartSlaveCmd);
         this.launchTimeoutSeconds = null;
@@ -484,8 +433,8 @@ public class SSHLauncher extends ComputerLauncher {
      * @param port       The port to connect on.
      * @param credentials The credentials to connect as.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
-     * @param jdkInstaller The jdk installer that will be used if no java vm is found on the specified host. If <code>null</code> the {@link DefaultJDKInstaller} will be used.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected.
+     * @param jdkInstaller not used.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      *                            @deprecated
@@ -504,8 +453,8 @@ public class SSHLauncher extends ComputerLauncher {
      * @param port       The port to connect on.
      * @param credentials The credentials to connect as.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
-     * @param jdkInstaller The jdk installer that will be used if no java vm is found on the specified host. If <code>null</code> the {@link DefaultJDKInstaller} will be used.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected .
+     * @param jdkInstaller not used.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      * @param launchTimeoutSeconds Launch timeout in seconds
@@ -542,8 +491,8 @@ public class SSHLauncher extends ComputerLauncher {
      * @param port       The port to connect on.
      * @param credentialsId The credentials to connect as.
      * @param jvmOptions Options passed to the java vm.
-     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected or installed by the JDKInstaller.
-     * @param jdkInstaller The jdk installer that will be used if no java vm is found on the specified host. If <code>null</code> the {@link DefaultJDKInstaller} will be used.
+     * @param javaPath   Path to the host jdk installation. If <code>null</code> the jdk will be auto detected.
+     * @param jdkInstaller not used.
      * @param prefixStartSlaveCmd This will prefix the start slave command. For instance if you want to execute the command with a different shell.
      * @param suffixStartSlaveCmd This will suffix the start slave command.
      * @param launchTimeoutSeconds Launch timeout in seconds
@@ -559,9 +508,6 @@ public class SSHLauncher extends ComputerLauncher {
         this.port = port == 0 ? 22 : port;
         this.credentialsId = credentialsId;
         this.javaPath = fixEmpty(javaPath);
-        if (jdkInstaller != null) {
-            this.jdk = jdkInstaller;
-        }
         this.prefixStartSlaveCmd = fixEmpty(prefixStartSlaveCmd);
         this.suffixStartSlaveCmd = fixEmpty(suffixStartSlaveCmd);
         this.launchTimeoutSeconds = launchTimeoutSeconds == null || launchTimeoutSeconds <= 0 ? null : launchTimeoutSeconds;
@@ -770,14 +716,6 @@ public class SSHLauncher extends ComputerLauncher {
     }
 
     /**
-     * Gets the optionnal java command to use to launch the slave JVM.
-     * @return
-     */
-    public String getJavaPath() {
-        return javaPath == null ? "" : javaPath;
-    }
-
-    /**
      * Gets the formatted current time stamp.
      *
      * @return the formatted current time stamp.
@@ -793,9 +731,10 @@ public class SSHLauncher extends ComputerLauncher {
      * @param computer The slave computer to get the root workspace of.
      *
      * @return the remote root workspace (without trailing slash).
+     *
      */
     @CheckForNull
-    private static String getWorkingDirectory(SlaveComputer computer) {
+    public static String getWorkingDirectory(SlaveComputer computer) {
         return getWorkingDirectory(computer.getNode());
     }
 
@@ -836,13 +775,21 @@ public class SSHLauncher extends ComputerLauncher {
                     verifyNoHeaderJunk(listener);
                     reportEnvironment(listener);
 
-                    String java = resolveJava(computer, listener);
-
                     final String workingDirectory = getWorkingDirectory(computer);
                     if (workingDirectory == null) {
                         listener.error("Cannot get the working directory for " + computer);
                         return Boolean.FALSE;
                     }
+
+                    String java = null;
+                    if (StringUtils.isNotBlank(javaPath)) {
+                        java = expandExpression(computer, javaPath);
+                    } else {
+                        JavaVersionChecker javaVersionChecker = new JavaVersionChecker(computer, listener, getJvmOptions(),
+                                                                                       connection);
+                        java = javaVersionChecker.resolveJava();
+                    }
+
                     copySlaveJar(listener, workingDirectory);
 
                     startSlave(computer, listener, java, workingDirectory);
@@ -918,44 +865,6 @@ public class SSHLauncher extends ComputerLauncher {
             connection = null;
             listener.getLogger().println(Messages.SSHLauncher_ConnectionClosed(getTimestamp()));
         }
-    }
-
-    /**
-     * return javaPath if specified in the configuration.
-     * Finds local Java, and if none exist, install one.
-     */
-    protected String resolveJava(SlaveComputer computer, TaskListener listener) throws InterruptedException, IOException {
-
-        if (StringUtils.isNotBlank(javaPath)) {
-            return expandExpression(computer, javaPath);
-        }
-
-        final String workingDirectory = getWorkingDirectory(computer);
-        if (workingDirectory == null) {
-            throw new IOException("Cannot retrieve a working directory of " + computer, null);
-        }
-
-        List<String> tried = new ArrayList<String>();
-        for (JavaProvider provider : JavaProvider.all()) {
-            for (String javaCommand : provider.getJavas(computer, listener, connection)) {
-                LOGGER.fine("Trying Java at "+javaCommand);
-                try {
-                    tried.add(javaCommand);
-                    return checkJavaVersion(listener, javaCommand);
-                } catch (IOException e) {
-                    LOGGER.log(FINE, "Failed to check the Java version",e);
-                    // try the next one
-                }
-            }
-        }
-        JDK javaJDK = Jenkins.getInstance().getJDK("java");
-        ExtensionList.lookup(JDK.DescriptorImpl.class).get(0)
-                     .getDefaultInstallers().get(0)
-                     .performInstallation(javaJDK, computer.getNode(), listener);
-
-        new JDKInstaller(SSHLauncher.DEFAULT_JDK, true);
-        throw new IOException("Java not found " + computer + ", install a Java 8+ version on the Agent, if you need "
-                              + "others JDK use the Global Tool Configuration to install them.");
     }
 
     private String expandExpression(SlaveComputer computer, String expression) {
@@ -1177,76 +1086,6 @@ public class SSHLauncher extends ComputerLauncher {
     protected void reportEnvironment(TaskListener listener) throws IOException, InterruptedException {
         listener.getLogger().println(Messages._SSHLauncher_RemoteUserEnvironment(getTimestamp()));
         connection.exec("set",listener.getLogger());
-    }
-
-    @NonNull
-    private String checkJavaVersion(TaskListener listener, String javaCommand) throws IOException, InterruptedException {
-        listener.getLogger().println(Messages.SSHLauncher_CheckingDefaultJava(getTimestamp(),javaCommand));
-        StringWriter output = new StringWriter();   // record output from Java
-
-        ByteArrayOutputStream out = new ByteArrayOutputStream();
-        connection.exec(javaCommand + " "+getJvmOptions() + " -version",out);
-        //TODO: Seems we need to retrieve the encoding from the connection destination
-        BufferedReader r = new BufferedReader(new InputStreamReader(
-                new ByteArrayInputStream(out.toByteArray()), Charset.defaultCharset()));
-        final String result = checkJavaVersion(listener.getLogger(), javaCommand, r, output);
-
-        if(null == result) {
-            listener.getLogger().println(Messages.SSHLauncher_UnknownJavaVersion(javaCommand));
-            listener.getLogger().println(output);
-            throw new IOException(Messages.SSHLauncher_UnknownJavaVersion(javaCommand));
-        } else {
-            return result;
-        }
-    }
-
-    // XXX switch to standard method in 1.479+
-    /**
-     * Given the output of "java -version" in <code>r</code>, determine if this
-     * version of Java is supported. This method has default visiblity for testing.
-     *
-     * @param logger
-     *            where to log the output
-     * @param javaCommand
-     *            the command executed, used for logging
-     * @param r
-     *            the output of "java -version"
-     * @param output
-     *            copy the data from <code>r</code> into this output buffer
-     */
-    @CheckForNull
-    protected String checkJavaVersion(final PrintStream logger, String javaCommand,
-            final BufferedReader r, final StringWriter output)
-            throws IOException {
-        String line;
-        while (null != (line = r.readLine())) {
-            output.write(line);
-            output.write("\n");
-            line = line.toLowerCase(Locale.ENGLISH);
-            if (line.startsWith("java version \"")
-                    || line.startsWith("openjdk version \"")) {
-                final String versionStr = line.substring(
-                        line.indexOf('\"') + 1, line.lastIndexOf('\"'));
-                logger.println(Messages.SSHLauncher_JavaVersionResult(
-                        getTimestamp(), javaCommand, versionStr));
-
-                // parse as a number and we should be OK as all we care about is up through the first dot.
-                final VersionNumber minJavaLevel = JavaProvider.getMinJavaLevel();
-                try {
-                    final Number version =
-                        NumberFormat.getNumberInstance(Locale.US).parse(versionStr);
-                    //TODO: burn it with fire
-                    if(version.doubleValue() < Double.parseDouble("1."+minJavaLevel)) {
-                        throw new IOException(Messages
-                                .SSHLauncher_NoJavaFound2(line, minJavaLevel.toString()));
-                    }
-                } catch(final ParseException e) {
-                    throw new IOException(Messages.SSHLauncher_NoJavaFound2(line, minJavaLevel));
-                }
-                return javaCommand;
-            }
-        }
-        return null;
     }
 
     protected void openConnection(final TaskListener listener, final SlaveComputer computer) throws IOException, InterruptedException {
@@ -1653,45 +1492,6 @@ public class SSHLauncher extends ComputerLauncher {
         }
 
 
-    }
-
-    @Extension
-    public static class DefaultJavaProvider extends JavaProvider {
-        @Override
-        public List<String> getJavas(SlaveComputer computer, TaskListener listener, Connection connection) {
-            List<String> javas = new ArrayList<String>(Arrays.asList(
-                    "java",
-                    "/usr/bin/java",
-                    "/usr/java/default/bin/java",
-                    "/usr/java/latest/bin/java",
-                    "/usr/local/bin/java",
-                    "/usr/local/java/bin/java")); // this is where we attempt to auto-install
-
-            String workingDirectory = getWorkingDirectory(computer);
-            if (workingDirectory != null) {
-                javas.add(workingDirectory + "/jdk/bin/java");
-            }
-               
-            final Node node = computer.getNode();
-            DescribableList<NodeProperty<?>,NodePropertyDescriptor> list = 
-                    node != null ? node.getNodeProperties() : null;
-            if (list != null) {
-                Descriptor jdk = Jenkins.getActiveInstance().getDescriptorByType(JDK.DescriptorImpl.class);
-                for (NodeProperty prop : list) {
-                    if (prop instanceof EnvironmentVariablesNodeProperty) {
-                        EnvVars env = ((EnvironmentVariablesNodeProperty)prop).getEnvVars();
-                        if (env != null && env.containsKey("JAVA_HOME"))
-                            javas.add(env.get("JAVA_HOME") + "/bin/java");
-                    }
-                    else if (prop instanceof ToolLocationNodeProperty) {
-                        for (ToolLocation tool : ((ToolLocationNodeProperty)prop).getLocations())
-                            if (tool.getType() == jdk)
-                                javas.add(tool.getHome() + "/bin/java");
-                    }
-                }
-            }
-            return javas;
-        }
     }
 
     private static final Logger LOGGER = Logger.getLogger(SSHLauncher.class.getName());
